Remove `sort` from the CampaignSubscriptions V2 endpoint
## TL;DR

Due to this field not working and performance issues supporting it today
it was decided to remove it from the official documentation.

## Background

`sort` and `direction` on this endpoint have been broken for a long
time. Once we found out, direction was determined fixable at this time.
Unfortunatly, sort was not.